WIP refactor into various packages
In particular, this is meant to cement the fact that, given how I've
more fully developed the Piece constructor in this commit (and have
gotten rid of 'setFrameOrigin'), I've made it so that at least the
Piece constructor can't be invoked from Main.java. I think I can even
move board into its own package.

WIP:
- When selecting a new piece, the old one simply disappears, making
way for the new one. While not correct, the behavior at least looks
consistent, and so I'm happy so far.

A comment was also corrected for a stray backslash.