Import project into an independent Gradle project
That is, I'm going to publish this on Github, but outside the harness
given to us by Hyperskill; the project can now be downloaded, built,
and run just like any ordinary Java project.

Gradle adds its own stuff to .gitignore, in particular the 'build/'
directory; so we keep its mention of said directory, while removing
the one we had initially when working on this project under
Hyperskill.

Also, regarding Org files: as these are mainly for jotting down
miscellaneous notes, I don't consider them part of the project, and so
the project ignores them for now.