Endow project with proper package prefix
This is based on advice given in the series "Working Class Java" on
YouTube, by Ryan Michael Kay, in particular where he recommends giving
even personal projects a "world applicable" package prefix that, for
example, identifies your work on app stores and such.

I've also changed the moniker "tetris" to "jtetris" whenever it
appears in a package name, to better reflect the name of the
application itself.

Link to relevant section of video:

https://www.youtube.com/watch?v=jupqgK3_K08&t=107s