Implement Z-piece template with test
This merely consists of adding the 'Z' directory as a project
resource. This directory contains the successive rotations of the 'Z'
piece.

The successive (counterclockwise) rotations of any piece should be
processed in lexicographic order of the corresponding filename. In
practice, this corresponds to the first rotation (for example, of 'Z')
corresponding to 'Z/1.txt', the second to 'Z/2.txt', and so on.

The test added in this commit (our first test!) checks whether the
code reading the piece from the filename maps to the correct
coordinates.