Introduce Rotation record
This replaces the older List<Coordinate> representation we had.

My sole concern right now is that, at some point, it may be convenient
to iterate over a Rotation record. I may well define a function to
export a list, iterate over that, and simply create a new
Rotation. Or, I could maket the record iterable. We'll see.