feat: give Coordinate and Rotation toString methods
The main purpose of giving these classes toString methods is to make
the Test Results output in IntelliJ easier to read, since we're now
using ParameterizedTest, which prints the various test parameters.

And so, we'd like this printout to be more readable, that is, not be
something like "Rotation[a=.... b=....]", and so on with the other
objects printed out.