refactor: rename rowOffset/columnOffset parameters to dr/dc
This is to distinguish them properly from the rowOffset and
columnOffset member fields of the Piece class. These fields represent
the concept of a Piece having a positional offset into the Board,
while dr and dc (in analogy to the more commonly seen 'dx' and 'dy')
represent the deltas of these values as the piece gets translated
through the board.