refactor: replace 'update' with erase/paintRotation
I would've liked for this to be 'erase/paintPiece', but doing this at
the Rotation level makes refactoring 'rotateCounterclockwise' easier.

It may well be ideal at some point to do away with CircularBuffer, and
figure out how to compute the next Rotation as a function of the
previous one. We'll see.