feat: add method marking filled rows for deletion
This involves introducing a new GridToken type, 'Cleared'. The idea is
that this GridToken type is only used temporarily: we'll later sweep
the board to eliminate the cleared cells.

In fact, it would be an error if any such cells remained after
sweeping/flushing the board (that is, subjecting all cells to
gravity.)