refactor: make rotation methods stateless
That is, we can now treat rotations just as we do translations. In
fact, a nice effect of this is that rotations commute with
translations.

The trick was to let CircularBuffer<T> keep the old rotation operation
as a stateful "shifting" operation, which the child Piece class then
uses to implement its own stateless version. The term "rotation" has
been replaced with "shift" in the tests wherever it was meaningful to
do so.

The rotation operation accepts an optional (via overloading) integer
argument specifying the number of rotations to perform.

Corresponding refactors in Board and the test are present in this
commit.